pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiki-genano"
version = "0.1.0"
description = "Preference-optimized generation of developable nanobodies across 65 epitope targets (mAbs 2026)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Radheesh Sharma Meda" },
  { name = "Jigar Doshi" },
  { name = "Eswar Iyer" },
  { name = "Shankar Shastry" },
  { name = "Venkatesh Mysore", email = "venkatesh@aikium.com" },
]
keywords = ["nanobody", "protein-language-model", "DPO", "GDPO", "developability", "VHH"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]

# Runtime dependencies are intentionally unpinned here; reproducible pins live
# in the provided Docker image (docker/Dockerfile).
dependencies = [
  "torch>=2.1",
  "transformers>=4.45",
  "datasets>=2.14",
  "accelerate>=0.27",
  "peft>=0.10",
  "trl>=0.8",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "numpy>=1.24",
  "pandas>=2.0",
  "tqdm>=4.65",
]

[project.optional-dependencies]
evaluation = [
  "fair-esm>=2.0",
  "biopython>=1.81",
  "matplotlib>=3.7",
  "seaborn>=0.13",
]

[project.scripts]
aiki-genano = "aiki_genano.cli.__main__:main"

[project.urls]
Homepage = "https://github.com/aikium-public/aiki-genano"

[tool.setuptools.packages.find]
include = ["aiki_genano*"]
exclude = ["data*", "results*", "checkpoints*", "benchmarks*", "docker*", "configs*"]
